+++ Cargo.toml
@@ -18,6 +18,8 @@ actix-web = "4.0.1"
 actix-files = "0.6.0"
 actix-multipart = "0.4.0"
 actix-http = "3.0.4"
+utoipa = { version = "1.1.0", features = ["actix_extras"] }
+utoipa-swagger-ui = { version = "1", features = ["actix-web"] }
 figlet-rs = "0.1.3"
 indicatif = "0.16.2"
 colored = "2.0.0"
