+++ src/routes/admin/registration_key.rs
@@ -5,11 +5,12 @@ use crate::{
     database::entity::registration_keys,
     models::{
         admin::registration_key::{RegistrationKeyData, RegistrationKeyParams},
-        MessageResponse, Page, Response,
+        MessageResponse, Page,
     },
     state::State,
     util::{
         auth::{auth_role, Auth},
+        response::Response,
         validate_paginate,
     },
 };
