+++ src/routes/mod.rs
@@ -1,8 +1,4 @@
-use crate::{
-    database::entity::settings,
-    models::{AppInfo, Response},
-    state::State,
-};
+use crate::{database::entity::settings, models::AppInfo, state::State, util::response::Response};
 use actix_web::{get, web, HttpResponse, Responder, Scope};
 use sea_orm::EntityTrait;
 
@@ -16,6 +12,13 @@ pub fn get_routes() -> Scope {
     web::scope("").service(info)
 }
 
+/// Get public server configuration
+#[utoipa::path(
+    context_path = "/api",
+    responses(
+        (status = 200, body = AppInfo)
+    )
+)]
 #[get("/info")]
 async fn info(state: web::Data<State>) -> Response<impl Responder> {
     let settings = settings::Entity::find_by_id(true)
